Refactor print camera property
