+++ include/rtweekend.h
@@ -36,27 +36,6 @@ constexpr double pi{ std::numbers::pi };
     return min + (max - min) * random_double();
 }
 
-// Camera
-
-// A utility function to print camera properties
-template <typename T>
-inline void print_camera_property_formatted(const std::string& property, const T& value) {
-    constexpr int format_width_left{ 18 }; // Length of "Samples per pixel" + 1
-    constexpr int format_width_right{ 4 };
-    std::cout << std::left << std::setw(format_width_left) << property
-                << std::right << std::setw(format_width_right) << value << "\n";
-}
-
-// Overload for Vec3
-inline void print_camera_property_formatted(const std::string& property, const Vec3& v) {
-    constexpr int format_width_left{ 15 };
-    constexpr int format_width_component{ 2 };
-    std::cout << std::left << std::setw(format_width_left) << property
-                << std::right << std::setw(format_width_component) << v.x() << ", "
-                << std::right << std::setw(format_width_component) << v.y() << ", "
-                << std::right << std::setw(format_width_component) << v.z() << "\n";
-}
-
 } // namespace rt
 
 #endif // INCLUDE_RTWEEKEND_H_
